Refactor data bindings
