+++ Include/RmlUi/Core.h
@@ -39,7 +39,9 @@
 #include "Core/ComputedValues.h"
 #include "Core/Context.h"
 #include "Core/ContextInstancer.h"
-#include "Core/DataBinding.h"
+#include "Core/DataController.h"
+#include "Core/DataModel.h"
+#include "Core/DataView.h"
 #include "Core/Decorator.h"
 #include "Core/DecoratorInstancer.h"
 #include "Core/Element.h"
