+++ Source/Core/DataView.cpp
@@ -27,8 +27,9 @@
  */
 
 #include "precompiled.h"
-#include "../../Include/RmlUi/Core/DataBinding.h"
-#include "../../Include/RmlUi/Core/Element.h"
+#include "../../Include/RmlUi/Core/DataView.h"
+#include "../../Include/RmlUi/Core/DataModel.h"
+
 
 namespace Rml {
 namespace Core {
@@ -111,7 +112,7 @@ bool DataViewText::Update(const DataModel& model)
 		}
 		else
 		{
-			Log::Message(Log::LT_WARNING, "Could not update data view text, parent element no longer valid. Was it destroyed?");
+			Log::Message(Log::LT_WARNING, "Could not update data view text, element no longer valid. Was it destroyed?");
 		}
 	}
 
@@ -177,12 +178,9 @@ DataViewIf::DataViewIf(const DataModel& model, Element* element, const String& b
 bool DataViewIf::Update(const DataModel& model)
 {
 	bool result = false;
-	Variant variant;
-	int int_value = 0;
-
-	if (model.GetValue(binding_name, variant) && variant.GetInto(int_value))
+	bool value = false;
+	if (model.GetValue(binding_name, value))
 	{
-		bool value = (int_value != 0);
 		bool is_visible = (element->GetLocalStyleProperties().count(PropertyId::Display) == 0);
 		if(is_visible != value)
 		{
@@ -193,129 +191,6 @@ bool DataViewIf::Update(const DataModel& model)
 			result = true;
 		}
 	}
-	return result;
-}
-
-DataControllerAttribute::DataControllerAttribute(const DataModel& model, const String& in_attribute_name, const String& in_value_name) : attribute_name(in_attribute_name), value_name(in_value_name)
-{
-	if (!model.IsWritable(value_name))
-	{
-		attribute_name.clear();
-		value_name.clear();
-	}
-}
-
-bool DataControllerAttribute::Update(Element* element, const DataModel& model) 
-{
-	bool result = false;
-	if (dirty)
-	{
-		if(Variant* value = element->GetAttribute(attribute_name))
-			result = model.SetValue(value_name, *value);
-		dirty = false;
-	}
-	return result;
-}
-
-
-bool DataModel::GetValue(const String& in_name, Variant& out_value) const
-{
-	bool success = true;
-
-	String name = in_name;
-	String member;
-
-	size_t i_dot = name.find('.');
-	if (i_dot != String::npos)
-	{
-		name = in_name.substr(0, i_dot);
-		member = in_name.substr(i_dot + 1);
-	}
-
-	auto it = bindings.find(name);
-	if (it != bindings.end())
-	{
-		const Binding& binding = it->second;
-
-		if (binding.type == ValueType::String)
-			out_value = *static_cast<const String*>(binding.ptr);
-		else if (binding.type == ValueType::Int)
-			out_value = *static_cast<const int*>(binding.ptr);
-		else if (binding.type == ValueType::Type)
-		{
-			success = false;
-			auto it_type = data_types.find(binding.data_type_name);
-			if(it_type != data_types.end())
-			{
-				const auto& members = it_type->second;
-				auto it_member = members.find(member);
-				if (it_member != members.end())
-				{
-					auto member_getset_ptr = it_member->second.get();
-					RMLUI_ASSERT(member_getset_ptr);
-					success = member_getset_ptr->Get(binding.ptr, out_value);
-				}
-			}
-			if(!success)
-				Log::Message(Log::LT_WARNING, "Could not get value from member '%s' in value named '%s' in data model.", member.c_str(), name.c_str());
-		}
-		else
-		{
-			RMLUI_ERRORMSG("TODO: Implementation for the provided binding type has not been made yet.");
-			success = false;
-		}
-	}
-	else
-	{
-		Log::Message(Log::LT_WARNING, "Could not find value named '%s' in data model.", name.c_str());
-		success = false;
-	}
-
-	return success;
-}
-
-
-bool DataModel::SetValue(const String& name, const Variant& value) const
-{
-	bool result = true;
-
-	auto it = bindings.find(name);
-	if (it != bindings.end())
-	{
-		const Binding& binding = it->second;
-
-		if (binding.writable)
-		{
-			if (binding.type == ValueType::String)
-				result = value.GetInto(*static_cast<String*>(binding.ptr));
-			else if (binding.type == ValueType::Int)
-				result = value.GetInto(*static_cast<int*>(binding.ptr));
-			else
-			{
-				RMLUI_ERRORMSG("TODO: Implementation for the provided binding type has not been made yet.");
-				result = false;
-			}
-		}
-		else
-		{
-			RMLUI_ERRORMSG("Controller attempted to write to a non-writable binding.");
-			result = false;
-		}
-	}
-	else
-	{
-		Log::Message(Log::LT_WARNING, "Could not find value named '%s' in data model.", name.c_str());
-		result = false;
-	}
-	return false;
-}
-
-bool DataModel::IsWritable(const String& name) const
-{
-	bool result = false;
-	auto it = bindings.find(name);
-	if (it != bindings.end())
-		result = it->second.writable;
 	return result;
 }
